Create new file for Turn class, create empty spec file
