+++ lib/board.rb
@@ -1,3 +1,5 @@
+require 'turn'
+
 class Board
   attr_reader :turns, :secret_code
 
@@ -54,23 +56,5 @@ private
     end
     board_string << "\n"
     board_string
-  end
-end
-
-class Turn
-  attr_accessor :feedback
-  attr_reader :guess
-
-  def self.with(guess)
-    Turn.new(guess)
-  end
-
-  def initialize(guess)
-    @guess = guess
-    @feedback = []
-  end
-
-  def to_s
-    "Black Pins: #{@feedback[0]} | White Pins: #{@feedback[1]} | Guess: #{@guess}\n"
   end
 end
